place PD erorrs in a separate file
